Fix issue where modals showing underneath WP Admin bar
Fixes issue where the excludes color box modal was underneath the WP
Admin bar. See https://cloudup.com/iyWDvvgXGWm

Promotes a modal across the plugin to be 85% of the height of the
window. Also modified the scrolling of the file lists as I found it
more difficult to scroll the modal and scroll the small box inside as
well.