Switch to using `wp_normalize_path` over `Backup::conform_dir` and remove the latter.
Unit tests have been left in for this commit to prove that they all still pass, with the exception of `testConformDirTestCase::testTrailingSlash`.

The behavious of `wp_normalize_path` differs from `conform_dir` when it comes to trailing slashes. We removed them but core leaves them be, thus our unit test for a path with a trailing slash fails. We're **not** actually relying on that behaviour anywhere so it doesn't actually cause any issues. Also looking into convention it seems that paths are generally better off having a trailing slash rather than not; for two reasons:

- a trailing slash is a nice indicator that a path is a directory is a path vs a file.
- having a double slashes in a path doesn't cause an error whereas ommitting the slash fully obviously does.

So we should switch to `wp_normalize_path` as is and remove our unit tests completely.