Beginnings of a welcome process
Introduces the conecept of a first-run process.

This will be an entirely new page, modeled on the core about page

The first run tasks will be broken into steps, each step represents something the user should do on the first run. Steps are marked with a tick when they are complete and are worked through in order, steps are collapsed showing only the title and icon, the current step is then expanded.

The whole process can be skipped which just takes people straight to the backups page.

When it's completed the user is taken to the backups page.