Avoids showing obsolete warnings when a backup succeeds.
If a backup_engine fails, triggers errors or warnings and then
a later fallback engine succeeds we previously surfaced those errors to the user.
This was confusing and could cause the user to think there was an issue when there
really wasn't.

This fixes that issue, we now clear errors and warnings from any previous backup_engines
as soon as one succeeds. We still show errors from all engines if they all fail.

Fixes https://github.com/humanmade/backupwordpress/issues/970