Ensure only new persistant notices are added added to the database
Previously, any new admin notices were merged into the array of persistant notices in the database without checking if the new notice already existed as a persistant notice, causing the database to fill up with multiple instances of the same message. This wasn't seen in the admin, because array_unique is used before outputting each notice.