Fix wp-snapshots
This wildcard set is causing wp-snapshots to NOT be excluded per report from user in forum. I confirmed this on my test site, and changed the plugin back to the previous iteration which was just using 'wp-snapshots' - and that worked perfectly. Unsure exactly why this happened but pretty sure the wildcard specification needs further translating for regex to do it correctly?